requirements.txt
flask
scrapy==1.1.0
service_identity
sqlalchemy
requests
scrapyd==1.1.0
scrapyd-client==1.0.1
psycopg2
python-dateutil
numpy
lxml
pyquery
nose
coverage
Cython>=0.21.1
lxml
scipy
mozsci
parsedatetime==2.1
pyopenssl==16.1.0
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pyvirtualdisplay